Add files via upload
background task will look for all transactions from NFT contract that get sent to a particular wallet, in this case Dead. It will the record those transactions into the burned table of the SQL database, following this the bot will then post an embed message to the discord server. The script will look every 30 minutes for any new entries and post them with no duplicates. If a manual execution it will post all entries in the burned database that has the value of NULL in the sent column.